sched/fair: Commit to lag based placement
Removes the FAIR_SLEEPERS code in favour of the new LAG based
placement.

Specifically, the whole FAIR_SLEEPER thing was a very crude
approximation to make up for the lack of lag based placement,
specifically the 'service owed' part. This is important for things
like 'starve' and 'hackbench'.

One side effect of FAIR_SLEEPER is that it caused 'small' unfairness,
specifically, by always ignoring up-to 'thresh' sleeptime it would
have a 50%/50% time distribution for a 50% sleeper vs a 100% runner,
while strictly speaking this should (of course) result in a 33%/67%
split (as CFS will also do if the sleep period exceeds 'thresh').

BUG=b:308209790
TEST=Manual tested on DUT

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lore.kernel.org/r/20230531124604.000198861@infradead.org
Signed-off-by: Youssef Esmat <youssefesmat@chromium.org>

Change-Id: I2659a4f476b0126c1a67df6ca952f5af07db223a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5089896
Tested-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>